deepfakemodel/requirements.txt
# Core Deep Learning
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Computer Vision
opencv-python>=4.5.0
pillow>=8.0.0
facenet-pytorch>=2.5.3

# Scientific Computing
numpy>=1.19.0
scipy>=1.5.0

# Machine Learning & Metrics
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Utilities
tqdm>=4.50.0

# Optional: For Jupyter notebook support
jupyter>=1.0.0
ipython>=7.0.0
